refactor: update file names
BREAKING CHANGE: rename `schema.ts` -> `schemas.ts`
BREAKING CHANGE: rename `server.ts` -> `servers.ts`
BREAKING CHANGE: rename `node.ts` -> `graph.ts`
BREAKING CHANGE: move `http.IExample` to `graph.INodeExample`
